Remove wrapper argument in AppWorld.msg
This commit removes the `AppWorldWrapper<W>` argument from the
`AppWorld.msg` trait.

We originally added this so that users could pass the wrapper into async
functions so that messages could be sent to the world from async
functions.

Instead, users should just store some sort of message sender (i.e. a
Tokio channel) that they can use to control how messages are sent to the
world.